[parse::animated] Move structures into their own file
